+++ scripts/data.js
@@ -3,7 +3,7 @@ export let noteList = {
         id: 111,
         cathegory: "Task",
         name: "shopping",
-        content: "buy: pasta, oil, bread",
+        content: "buy: pasta, oil, bread. 11/10/2021",
         created: "07/10/2021",
         dates: "11/10/2021",
     },
@@ -21,7 +21,7 @@ export let noteList = {
         name: "Project",
         content: "use CSS Grid in my case",
         created: "06/10/2021",
-        dates: "08/10/2021",
+        dates: "",
     },
     444: {
         id: 444,
@@ -51,7 +51,7 @@ export let noteList = {
         id: 777,
         cathegory: "Task",
         name: "Cleaning",
-        content: "Clean whole the flat",
+        content: "Clean whole the flat. Start on 10/10/2021",
         created: "09/10/2021",
         dates: "10/10/2021",
     },
@@ -84,24 +84,4 @@ export const addArchivedNote = (note) => {
         ...noteArchive,
         [note.id]: note,
     }
-}
-
-export const btnDone = (id) => {
-    return `<div class="done btn"><img id="${id}-done" src="icons/check.svg" alt="archive" title="Archive"></div>`;
-}
-
-export const btnEdit = (id) => {
-    return `<div class="edit btn"><img id="${id}-edit" src="icons/pencil.svg" alt="edit" title="Edit"></div>`;
-}
-
-export const btnTrash = (id) => {
-    return `<div class="trash btn"><img id="${id}-trash" src="icons/trash.svg" alt="delete" title="Delete"></div>`;
-}
-
-export const btnSave = (id) => {
-    return `<div class="save btn"><img id="${id}-save" src="icons/save.svg" alt="save" title="Save"></div>`;
-}
-
-export const btnCancel = (id) => {
-    return `<div class="cancel btn"><img id="${id}-cancel" src="icons/cancel.svg" alt="cancel" title="Cancel"></div>`;
 }
